perf(azurerm/clouds): do not search for instance on create but use single request (#82)
* perf(azurerm/clouds): do not search for instance on create but use single request

Instead of listing all instances and finding one matching the name, we make a single get request to fetch the instance directly. As we know in which resource group we created it.

* perf(azurerm/clouds): do not search for instance on create but use single request

Instead of listing all instances and finding one matching the name, we make a single get request to fetch the instance directly. As we know in which resource group we created it.

* style: typo

* fix(azurerm/cloud): remove client caching